requirements.txt
requests
fastapi
pyjwt
python-dotenv